+++ CHANGES.rst
@@ -5,6 +5,8 @@ Changelog
 ------------------
 
 - #1994 Support for dynamic assignment of Local Roles for context and principal
+- #1992 Fix Generic Setup XML export/import adapters for Dexterity fields
+- #1993 Avoid line wrapping in login form for small screens
 - #1990 Fix items not filtered by Worksheet Template's method in Add analyses
 - #1991 Update default worksheet layout 
 - #1887 Fix instruments not filtered by method in Worksheet Template edit view